extracting hybrid publisher
