+++ library/plugins/hybrid/publisher/build.gradle.kts
@@ -1,0 +1,8 @@
+plugins {
+    id("kotlinLib")
+}
+
+dependencies {
+    implementation(project(":library:talaiot"))
+    testImplementation(project(":library:talaiot-test-utils"))
+}
