+++ library/plugins/hybrid/publisher/src/main/kotlin/com/cdsap/talaiot/publisher/HybridPublisher.kt
