+++ library/plugins/talaiot/build.gradle.kts
@@ -14,6 +14,7 @@ talaiotPlugin {
 dependencies {
     implementation(project(":library:talaiot"))
     implementation(project(":library:plugins:base:publisher"))
+    //implementation(project(":library:plugins:hybrid:publisher"))
     implementation("guru.nidi:graphviz-java:0.8.3")
     implementation("org.influxdb:influxdb-java:2.19")
     implementation("com.github.oshi:oshi-core:3.13.3")
