+++ settings.gradle.kts
@@ -16,3 +16,4 @@ include(":library:talaiot-test-utils")
 include(":library:plugins:talaiot")
 include(":library:plugins:base:plugin")
 include(":library:plugins:base:publisher")
+include(":library:plugins:hybrid:publisher")
